Revert "No error, No warning"
This reverts commit 5d65d6b36c38ef8891a71e85636d82186a41714f.

+++ TreadMill/TreadMill/encoder.h
@@ -8,6 +8,6 @@
 #endif
 
 void encoder_init(void);
-uint8_t encoder_read(void);
+uint32_t encoder_read(void);
 
 #endif
